requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
uWSGI==2.0.22
Werkzeug==2.3.7
WTForms==3.0.1
